src/main/kotlin/com/yzzzzun/board/service/LikeService.kt
package com.yzzzzun.board.service

import com.yzzzzun.board.event.dto.LikeEvent
import com.yzzzzun.board.repository.LikeRepository
import com.yzzzzun.board.util.RedisUtil
import org.springframework.context.ApplicationEventPublisher
import org.springframework.stereotype.Service
import org.springframework.transaction.annotation.Transactional

@Service
@Transactional(readOnly = true)
class LikeService(
    private val likeRepository: LikeRepository,
    private val redisUtil: RedisUtil,
    private val applicationEventPublisher: ApplicationEventPublisher,
) {
    fun createLike(
        postId: Long,
        createdBy: String,
    ) {
        applicationEventPublisher.publishEvent(LikeEvent(postId, createdBy))
    }

    fun countLike(postId: Long): Long {
        redisUtil.getCount(redisUtil.getLikeCountKey(postId))?.let {
            return it
        }
        with(likeRepository.countByPostId(postId)) {
            redisUtil.setData(redisUtil.getLikeCountKey(postId), 1L)
            return this
        }
    }
}
